+++ apps/api/src/services/data-provider/financial-modeling-prep/financial-modeling-prep.service.ts
@@ -54,12 +54,15 @@ export class FinancialModelingPrepService
   implements DataProviderInterface, OnModuleInit
 {
   private static countriesMapping = {
+    'Congo (Dem. Rep. of the)': 'CD',
+    'Congo (Rep. of)': 'CG',
     'Czech Republic': 'CZ',
     'Korea (the Republic of)': 'KR',
     Macau: 'MO',
     'Russian Federation': 'RU',
     'Taiwan (Province of China)': 'TW',
-    Turkey: 'TR'
+    Turkey: 'TR',
+    'Virgin Islands (British)': 'VG'
   };
 
   private readonly logger = new Logger(FinancialModelingPrepService.name);